Update datacube_tools.py
Fixed an issue that arose when a user specified a point that was exactly on the boundary of a datacube - the point shift intended to deal with curvature differences in projections would not shift this point, so it continued to fail.  Now shifts point one meter off the boundary into the original cube found.